+++ src/vs/workbench/contrib/void/electron-main/voidSCMMainService.ts
